requirements.txt
PyJWT==2.8.0
cryptography==42.0.5
pyyaml==6.0.1
fastapi==0.110.0
uvicorn==0.27.1
lark-oapi==1.5.5
python-multipart==0.0.9
jieba==0.42.1
requests==2.31.0
openai>=1.0.0
jieba
bs4